apps/control-api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
websockets==13.1
pydantic==2.9.2
kafka-python==2.0.2
redis==5.1.1
psycopg2-binary==2.9.10
python-multipart==0.0.12
prometheus_client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-prometheus==0.48b0
colorama>=0.4.6
